Added support for extra dependencies in the params.pp manifest.
The parameters' manifest now support optional extra dependencies. These
are passed in by the client through Hiera. Each comes in the form of a
hash, whose value and key are, respectively, the pack name and the
provider.

These extra dependencies will be then merged with the mandatory ones
into a unique set of packages that must be installed.